ARPA-541 Prevent last admin from being deleted

+++ Orso.Arpa.Domain/Logic/Users/Delete.cs
@@ -1,10 +1,16 @@
+using System;
+using System.Collections.Generic;
+using System.Linq;
 using System.Threading;
 using System.Threading.Tasks;
+using FluentValidation;
+using FluentValidation.Results;
 using MediatR;
 using Microsoft.AspNetCore.Identity;
 using Orso.Arpa.Domain.Entities;
 using Orso.Arpa.Domain.Errors;
 using Orso.Arpa.Domain.Identity;
+using Orso.Arpa.Domain.Roles;
 
 namespace Orso.Arpa.Domain.Logic.Users
 {
@@ -38,6 +44,11 @@ namespace Orso.Arpa.Domain.Logic.Users
                     throw new NotFoundException(nameof(User), nameof(Command.UserName));
                 }
 
+                if (await CheckIfLastAdminWillBeRemovedAsync(user.Id))
+                {
+                    throw new ValidationException(new ValidationFailure[] { new ValidationFailure(nameof(request.UserName), "The operation is not allowed because it would remove the last administrator") });
+                }
+
                 IdentityResult result = await _userManager.DeleteAsync(user);
 
                 if (result.Succeeded)
@@ -45,6 +56,12 @@ namespace Orso.Arpa.Domain.Logic.Users
                     return Unit.Value;
                 }
                 throw new IdentityException("Problem deleting user", result.Errors);
+            }
+
+            private async Task<bool> CheckIfLastAdminWillBeRemovedAsync(Guid userId)
+            {
+                IList<User> adminUsers = await _userManager.GetUsersInRoleAsync(RoleNames.Admin);
+                return adminUsers.Count == 1 && adminUsers.First().Id.Equals(userId);
             }
         }
     }
